fix dependencies version resolution to use the dependent package.json (#3524)
The issue started once the resolveNodePackage has changed to return an object with the package-json data. The way how to get the version from that data is different.
This PR fetches the data correctly according to the ResolvedNodePackage interface. It first tries to use the concrete installed version in the node-modules and if not found, it uses the version written in the dependent package-json file.